test: 볼카운트 테스트
